Move Customization to separate project
Allows plugins to depend directly on customization rather than the
whole of SCCMCliCtrWPF, meaning it can then reference plugins to
automatically include them in builds.

The new project is placed in the root of the repo as there are
existing references to a Customization project there, eg several
plugins already look there for the DLL, so no changes are required
for them to function.